Add files via upload
Changed how actor's shoot bullets, they now do it through a weapons class whose constructor requires a player owner and a fire rate. This allows for easily variable fire rate and opens the door for more implementations of shooting. It also allows shooting to be in the player's fastTimer as the speed of a player's action handling will no longer have an effect on the speed at which a player can fire. 

Implemented limitations on the movement of actors, they can no longer move off the screen. At the moment there is a buffer between the edge and the player everywhere but on the top border. This is not hard coded and this is not a bug. This is a function of the way the player move's (increments of ten) and the width of both the player and the screen. Smaller movements would solve this as the player would be able to get closer to the edges. We ought to implement this change once the player and their movement are made scalable with screen size. 

Implemented getHeight() and getWidth() metho