Add files via upload
Made default testing resolution 1920 x 1080 pxl. Added distortion overloads to account for changed scale and to allow actor objects to have a distortion factor generated by state as a parameter in their constructor. Also changed Qt::Key design in player to allow for settings to eventually be modified and for two valid keys to exist for certain actions such as movement.